Divided the app into main file and a UI file to better the readibility of the code for later
